test: move recommendation database retrieval coverage to unit tests
Reclassify unit-level coverage from tests/test_containers_integration_tests/services/recommend_app/test_database_retrieval.py into the unit test suite, or remove it when equivalent unit coverage already exists. Preserve real container-backed coverage and use SQLite sessions for database-dependent unit paths.
